% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialCRIs.R
\name{binomial_pi}
\alias{binomial_pi}
\title{Calculates the percentile interval for the probability parameter
underlying a binomially distributed outcome.}
\usage{
binomial_pi(n_successes, n_trials, prob, prior_shape1 = 1, prior_shape2 = 1)
}
\arguments{
\item{n_successes}{The number of successes.}

\item{n_trials}{The total number of trials.}

\item{prob}{The size of the percentile interval with 0 <= prob <= 1.}

\item{prior_shape1}{The shape1 parameter of the Beta distribution
defining the prior.  The default values shape1=1 and shape2=1 define
a flat prior assigning equal probability density to all possible
parameter values.}

\item{prior_shape2}{The shape2 parameter of the Beta distribution defining
the prior.  The default values shape1=1 and shape2=1 define a flat
prior assigning equal probability density to all possible
parameter values.}
}
\value{
A vector containing the endpoints of the percentile interval.
}
\description{
Calculates the percentile interval for the probability parameter
underlying a binomially distributed outcome.
}
\examples{
# 6/9 successes with flat prior:
binomial_pi(6, 9, 0.8)

# 6/9 successes with prior assuming one prior success and one failure:
binomial_pi(6, 9, 0.8, 2, 2)

# Equivalently:
binomial_pi(7, 11, 0.8)
}
\author{
Titus von der Malsburg <malsburg@uni-potsdam.de>
}
