% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialCRIs.R
\docType{package}
\name{binomialCRIs}
\alias{binomialCRIs}
\alias{binomialCRIs-package}
\title{Calculate and Plot Binomial Credible Intervals}
\description{
Given data and a prior, the functions in the package calculate and
plot credible intervals for the probability of success.
}
\seealso{
\code{\link{binomial_hpdi}}, \code{\link{binomial_pi}}, \code{\link{binomial_prob}}, \code{\link{plot_binomial_hpdi}}, \code{\link{plot_binomial_pi}}, \code{\link{plot_binomial_cri}}
}
\author{
Titus von der Malsburg \email{malsburg@uni-potsdam.de}
}
\keyword{Bayesian}
\keyword{binomial}
\keyword{credible}
\keyword{distribution,}
\keyword{inference,}
\keyword{intervals}
