% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialCRIs.R
\name{plot_binomial_pi}
\alias{plot_binomial_pi}
\title{Plots the posterior distribution and specified percentile interval
for the probability parameter underlying a binomially distributed
outcome.}
\usage{
plot_binomial_pi(n_successes, n_trials, prob, prior_shape1 = 1,
  prior_shape2 = 1)
}
\arguments{
\item{n_successes}{The number of successes.}

\item{n_trials}{The total number of trials.}

\item{prob}{The size of the percentile interval with 0 <= prob <= 1.}

\item{prior_shape1}{The shape1 parameter of the Beta distribution
defining the prior.  The default values shape1=1 and shape2=1 define
a flat prior assigning equal probability density to all possible
parameter values.}

\item{prior_shape2}{The shape2 parameter of the Beta distribution defining
the prior.  The default values shape1=1 and shape2=1 define a flat
prior assigning equal probability density to all possible
parameter values.}
}
\value{
A ggplot2 object.
}
\description{
Plots the posterior distribution and specified percentile interval
for the probability parameter underlying a binomially distributed
outcome.
}
\examples{
plot_binomial_pi(6, 9, 0.5)
}
\author{
Titus von der Malsburg <malsburg@uni-potsdam.de>
}
